вычитка перед защитой
