+++ js/main.js
@@ -5,4 +5,5 @@ import './card.js';
 import './map.js';
 import './form.js';
 import './avatar.js';
+import './filter.js';
